namespaced code to cpp classes

+++ src/const.h
@@ -7,7 +7,7 @@ namespace CONST {
     const unsigned int GAME_DEVICE_ID         = 3; // 1, 2, 3
     const unsigned int GAME_RADIUSES[]        = {305, 511, 362};
     
-    const unsigned int GAME_TEAMS_COUNT       = 5;
+    const byte         GAME_TEAMS_COUNT       = 5;
     const          int GAME_TEAM_RAW_VALUES[] = {
         415, //  15kOhm idx = 0 # SUPER 
         205, //  39kOhm idx = 1
